refactor(wallet): :recycle:  best server.ts organization
